Fixed handling of \part command, fixed references to Appendices in supplement.
I changed the labmanual.cls file to redefine how the \part command is
handled.  Now we no longer need to use the \newclude package.  That's
good, because \newclude is incompatible with the xr package, and we need
the xr package to be able to reference the appendices in the
master_supplement.tex file.