Changed fonts
I changed the font to "latin modern" instead of the default "computer
modern" (cm).   In print, the change is almost indistinguishable.  But
onscreen, at high zoom, the cm font appears pixelated, because it is
bitmapped instead of vectorized.  Apparently it's weird that the Miktex
distribution uses cm as a default.  Latin Modern is derived from cm, and
is generally preferred.  This change affects 131 and 132.  I've checked
all of the labs that were included this year, and made sure that this
change did not cause any terrible page breaks or anything.