Added option for each lab to start on odd page
If each lab starts on an odd page, it's easier to have students pull out
(and hand in) a single lab at a time; no lab starts on the back side of
another.  Tradeoff is that lab manual is longer by about 16 pages, so
it's 8 sheets heavier and about $1 more expensive.  The option is
selectable by commenting or uncommenting a single line in master.tex