Finished implementing changes to various figures
Mariama made minor changes to various figures over the last few days, but didn't have a convenient way to convert docx and pptx files to pdf and eps, or to crop them.  I've completed it here.  (Note that I have NOT yet done the figures from the elec_grav lab.