Implemented figures recently changed by Mariama
Over the last day or two, Mariama has made many changes to the underlying docx and pptx files for a bunch of figures.  (I disagreed with a few of those changes by email, and we've settled on not changing the fonts in most of the figures, for now; other changes in notation are fine.)  With this commit, I THINK I have implemented all of the changes to the figures (from microsoft documents to cropped eps files),  changing filenames in the tex files as necessary.  (Keeping things straight was a little complicated, partly because I apparently forgot to push some earlier changes yesterday; sorry about that.  Because of that, I suppose it's possible I've missed something here.)

I think the one change I did NOT fully implement was the three colored arrows in figure 1 of the Forces and Motion II lab.  I used the new figure, but redid all arrows in black, mainly since we're not changing all of the other arrows to color yet anyway.  In fact, at the current price of +$0.40 for each color page, I strongly think that colored arrows are not worth the cost, and I request that we put a hold on any new color pages until we can all discuss this together this fall.

Thanks for your work on all the figures, Mariama!  :-)