Fixed figure 4 in newtons_laws
This is complicated.  I think Mariama had earlier tweaked the pptx file, but also provided a non-tweaked .eps file that was actually possibly corrupted (large, and crashed some pdf viewers).  This version implements the tweaked notation, but in an eps that was created the same as all others in this manual.  Should be okay now.