New lab on normal force in elevators
This is a lab I've done as an in-class exercise with students for years, but I've written it up here as a formal worksheet.  I ran it in class this year, and I found that making them draw the graphs focuses them better than when I lead them through it.
I have also included the lab in the master, and also in the (new!) spreadsheet for spring 2020.