Initial framework for endnotes (for instructor notes)
This is an initial commit to set up the framework we would need to handle endnotes, which we could use to create an optional "Instructors notes" at the end of the manual (not for students).  It sounds easy enough, but in fact there is a lot of stuff to set up, because we need to make the endnotes play nicely with all of the other weird stuff we're doing in the manual with sections, parts, pagebreaks, footnotes, headers, the table of contents, and everything else.  For now, I've added just a few "dummy" notes to two 131 labs.  Currently, the system is workable as a proof of concept, but it's kinda ugly. (The header for the notes section doesn't work, for instance.)