New version of pendulum lab
This new version uses the pasco photogate to time the period of the pendulum, which allows a timing accurate enough to see the small effect that the amplitude has on the period for large oscillations.  There were also a few other aspects of the old version that I wasn't excited about, so I wrote this new version rather than tacking on a new section to the old version.