created batch file for compiling files for Phys131 directory
The general plan here is that individual .cap files (or notebook files, or movies, or whatever we need) will be included in the directories that hold each individual lab.  The batch file make_phys131.bat compiles all of the files listed in file_list.txt into a single directory, Phys131, which can then be copied to the computers in the classrooms.