Removed small type note
This is an old lab, no longer used in 131.  For whatever reason, the {\small } command started causing an error on my MikTeX distribution.  I was too lazy to fix the error, and we don't really want small text there anyway, so I just allowed the text to be normal sized.