Fixed problem with outline (bookmarks) in pdf output file
This was the weirdest thing.  The hyperref package was screwing up the outline (known as the "bookmarks") in the pdf output file.  Instead of showing an outline of 7 parts, it showed only 3, with the last 4 incorrectly placed at the next level down in the outline.  This problem only occurred under pdflatex, not under xelatex.  For whatever reason, including the bookmarks command fixed the problem.