Changes to geometry in labmanualf_formatting_commands
For a long time the lab manuals have used the geometry package, with what seemed to me like a really weird set of parameters--including a totally arbitrary offset that needed to be set when using at least some versions of MikTex.  I simplified it all down to just four lines, though for now I've kept the old parameters commented out.

The immediate reason for doing this was that I had started to use the package pdfpages for something else, and the kludgy hoffset and voffset messed that up.